requirements.txt
streamlit
datasets
python-dotenv
langchain==0.1.0
openai==1.7.2
langchain-openai==0.0.2
langchain-community==0.0.12
langchainhub==0.1.14
chromadb==0.4.22